# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bot-Mox API v1
  version: 1.0.0
  description: Canonical API surface for phased backend/frontend alignment.
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/bots:
    get:
      summary: List bots
      responses:
        '200':
          description: Bot list
    post:
      summary: Create bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created bot
  /api/v1/bots/{id}:
    get:
      summary: Get bot by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bot entity
    patch:
      summary: Update bot by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated bot
    delete:
      summary: Delete bot by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted bot
  /api/v1/bots/{id}/lifecycle:
    get:
      summary: Get bot lifecycle payload
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bot lifecycle
        '404':
          description: Not found
  /api/v1/bots/{id}/lifecycle/transitions:
    get:
      summary: Get bot lifecycle stage transitions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lifecycle transitions
  /api/v1/bots/{id}/lifecycle/is-banned:
    get:
      summary: Check whether bot is banned
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Banned state
  /api/v1/bots/{id}/lifecycle/transition:
    post:
      summary: Transition bot lifecycle status (non-ban)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [offline, prepare, leveling, profession, farming, banned]
              required: [status]
      responses:
        '200':
          description: Updated bot
  /api/v1/bots/{id}/lifecycle/ban:
    post:
      summary: Ban bot and persist ban details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ban_date:
                  type: string
                  example: 10.02.2026
                ban_reason:
                  type: string
                ban_mechanism:
                  type: string
              required: [ban_date, ban_reason, ban_mechanism]
      responses:
        '200':
          description: Bot banned
  /api/v1/bots/{id}/lifecycle/unban:
    post:
      summary: Remove bot ban and restore previous status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bot unbanned
  /api/v1/finance/operations:
    get:
      summary: List finance operations
      responses:
        '200':
          description: Finance operation list
    post:
      summary: Create finance operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [income, expense]
                category:
                  type: string
                amount:
                  type: number
                currency:
                  type: string
                date:
                  type: integer
              required: [type, category, amount, currency, date]
      responses:
        '201':
          description: Finance operation created
  /api/v1/finance/operations/{id}:
    get:
      summary: Get finance operation by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Finance operation
        '404':
          description: Not found
    patch:
      summary: Update finance operation by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated finance operation
    delete:
      summary: Delete finance operation by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
...